chore: rename folder `test` to `tests`
