+++ tests/integration/api/v1/sessions/delete.test.js
@@ -1,7 +1,7 @@
 import { version as uuidVersion } from "uuid";
 import setCookieParser from "set-cookie-parser";
 
-import orchestrator from "test/orchestrator";
+import orchestrator from "tests/orchestrator";
 import session from "models/session";
 
 beforeAll(async () => {
