+++ tests/integration/api/v1/status/get.test.js
@@ -1,4 +1,4 @@
-import orchestrator from "test/orchestrator";
+import orchestrator from "tests/orchestrator";
 
 beforeAll(async () => {
   await orchestrator.waitAllServices();
