+++ tests/integration/api/v1/users/[username]/get.test.js
@@ -1,5 +1,5 @@
 import { version as uuidVersion } from "uuid";
-import orchestrator from "test/orchestrator";
+import orchestrator from "tests/orchestrator";
 
 beforeAll(async () => {
   await orchestrator.waitAllServices();
