+++ tests/integration/infra/email.test.js
@@ -1,5 +1,5 @@
 import email from "infra/email.js";
-import orchestrator from "test/orchestrator.js";
+import orchestrator from "tests/orchestrator.js";
 
 beforeAll(async () => {
   await orchestrator.waitAllServices();
